+++ src/SmugMugModel.v2/Types/Entities/AlbumTemplate/AlbumTemplate.properties.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
@@ -30,7 +31,7 @@ namespace SmugMug.v2.Types
         private string _passwordHint;
         private bool _printable;
         private string _printmarkUri;
-        private int _proofDays;
+        private long _proofDays;
         private bool _protected;
         private bool _public;
         private bool _share;
@@ -366,7 +367,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int ProofDays {
+        public long ProofDays {
             get {
                 return _proofDays;
             }
@@ -532,8 +533,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
