+++ src/SmugMugModel.v2/Types/Entities/Comment/Comment.properties.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
@@ -10,9 +11,9 @@ namespace SmugMug.v2.Types
         private string _comment;
         private DateTime _date;
         private string _name;
-        private int _rating;
-        private int _serviceID;
-        private int _socialID;
+        private long _rating;
+        private long _serviceID;
+        private long _socialID;
         private CommentStatusEnum _status;
 
 
@@ -58,7 +59,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Rating {
+        public long Rating {
             get {
                 return _rating;
             }
@@ -72,7 +73,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int ServiceID {
+        public long ServiceID {
             get {
                 return _serviceID;
             }
@@ -86,7 +87,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int SocialID {
+        public long SocialID {
             get {
                 return _socialID;
             }
@@ -112,8 +113,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
