+++ src/SmugMugModel.v2/Types/Entities/Download/Download.properties.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
@@ -10,7 +11,7 @@ namespace SmugMug.v2.Types
         private DateTime _dateAdded;
         private DateTime _dateModified;
         private string _fileName;
-        private int _part;
+        private long _part;
         private ImageStatusEnum _status;
 
 
@@ -56,7 +57,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Part {
+        public long Part {
             get {
                 return _part;
             }
@@ -82,8 +83,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
