+++ src/SmugMugModel.v2/Types/Entities/ImageSizeCustom/ImageSizeCustom.properties.cs
@@ -2,14 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
     public partial class ImageSizeCustomEntity : SmugMugEntity
     {
         private string _ext;
-        private int _requestedHeight;
-        private int _requestedWidth;
+        private long _requestedHeight;
+        private long _requestedWidth;
         private string _url;
         private bool _usable;
         private bool _watermarked;
@@ -29,7 +30,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int RequestedHeight {
+        public long RequestedHeight {
             get {
                 return _requestedHeight;
             }
@@ -43,7 +44,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int RequestedWidth {
+        public long RequestedWidth {
             get {
                 return _requestedWidth;
             }
@@ -97,8 +98,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
