+++ src/SmugMugModel.v2/Types/Entities/LargestImage/LargestImage.properties.cs
@@ -2,20 +2,21 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
     public partial class LargestImageEntity : SmugMugEntity
     {
         private string _ext;
-        private int _height;
-        private int _size;
+        private long _height;
+        private long _size;
         private string _uri;
         private string _uriDescription;
         private string _url;
         private bool _usable;
         private bool _watermarked;
-        private int _width;
+        private long _width;
 
 
         public string Ext {
@@ -32,7 +33,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Height {
+        public long Height {
             get {
                 return _height;
             }
@@ -46,7 +47,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Size {
+        public long Size {
             get {
                 return _size;
             }
@@ -130,7 +131,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Width {
+        public long Width {
             get {
                 return _width;
             }
@@ -142,8 +143,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
