+++ src/SmugMugModel.v2/Types/Entities/LargestVideo/LargestVideo.properties.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
@@ -10,13 +11,13 @@ namespace SmugMug.v2.Types
         private decimal _bitrate;
         private decimal _duration;
         private string _ext;
-        private int _height;
-        private int _size;
+        private long _height;
+        private long _size;
         private string _uri;
         private string _uriDescription;
         private string _url;
         private bool _usable;
-        private int _width;
+        private long _width;
 
 
         public decimal Bitrate {
@@ -61,7 +62,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Height {
+        public long Height {
             get {
                 return _height;
             }
@@ -75,7 +76,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Size {
+        public long Size {
             get {
                 return _size;
             }
@@ -145,7 +146,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Width {
+        public long Width {
             get {
                 return _width;
             }
@@ -157,8 +158,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
