+++ src/SmugMugModel.v2/Types/Entities/Printmark/Printmark.properties.cs
@@ -2,23 +2,24 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
     public partial class PrintmarkEntity : SmugMugEntity
     {
-        private int _dissolve;
-        private int _gap;
-        private int _imageID;
+        private long _dissolve;
+        private long _gap;
+        private long _imageID;
         private string _imageKey;
         private string _imageUri;
         private string _location;
         private string _name;
-        private int _printmarkID;
-        private int _scale;
+        private long _printmarkID;
+        private long _scale;
 
 
-        public int Dissolve {
+        public long Dissolve {
             get {
                 return _dissolve;
             }
@@ -32,7 +33,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Gap {
+        public long Gap {
             get {
                 return _gap;
             }
@@ -46,7 +47,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int ImageID {
+        public long ImageID {
             get {
                 return _imageID;
             }
@@ -116,7 +117,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int PrintmarkID {
+        public long PrintmarkID {
             get {
                 return _printmarkID;
             }
@@ -130,7 +131,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Scale {
+        public long Scale {
             get {
                 return _scale;
             }
@@ -142,8 +143,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
