+++ src/SmugMugModel.v2/Types/Entities/StatusImageRotate/StatusImageRotate.properties.cs
@@ -2,22 +2,23 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
     public partial class StatusImageRotateEntity : SmugMugEntity
     {
-        private int _degrees;
+        private long _degrees;
         private string _imageUri;
         private DateTime _lastUpdated;
         private string _message;
         private string _queueName;
-        private int _rotate;
+        private long _rotate;
         private ImageStatusEnum _status;
         private DateTime _timeIn;
 
 
-        public int Degrees {
+        public long Degrees {
             get {
                 return _degrees;
             }
@@ -87,7 +88,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Rotate {
+        public long Rotate {
             get {
                 return _rotate;
             }
@@ -127,8 +128,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
