+++ src/SmugMugModel.v2/Types/Entities/Template/Template.properties.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
@@ -12,7 +13,7 @@ namespace SmugMug.v2.Types
         private bool _dark;
         private string _filename;
         private string _name;
-        private int _thumbs;
+        private long _thumbs;
 
 
         public bool AjaxEnabled {
@@ -85,7 +86,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Thumbs {
+        public long Thumbs {
             get {
                 return _thumbs;
             }
@@ -97,8 +98,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
