+++ src/SmugMugModel.v2/Types/Entities/VideoSMIL/VideoSMIL.properties.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
@@ -10,15 +11,15 @@ namespace SmugMug.v2.Types
         private decimal _bitrate;
         private decimal _duration;
         private string _ext;
-        private int _height;
+        private long _height;
         private string _mD5;
-        private int _size;
+        private long _size;
         private string _uri;
         private string _uriDescription;
         private string _url;
         private bool _usable;
         private bool _watermarked;
-        private int _width;
+        private long _width;
 
 
         public decimal Bitrate {
@@ -63,7 +64,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Height {
+        public long Height {
             get {
                 return _height;
             }
@@ -91,7 +92,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Size {
+        public long Size {
             get {
                 return _size;
             }
@@ -175,7 +176,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Width {
+        public long Width {
             get {
                 return _width;
             }
@@ -187,8 +188,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
