+++ src/SmugMugModel.v2/Types/Entities/Watermark/Watermark.properties.cs
@@ -2,22 +2,23 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using SmugMug.v2.Authentication;
 
 namespace SmugMug.v2.Types
 {
     public partial class WatermarkEntity : SmugMugEntity
     {
-        private int _dissolve;
-        private int _imageID;
+        private long _dissolve;
+        private long _imageID;
         private string _imageKey;
         private string _name;
         private PinnedEnum _pinned;
         private SmallestSizeEnum _smallestSize;
         private bool _thumbs;
-        private int _watermarkID;
+        private long _watermarkID;
 
 
-        public int Dissolve {
+        public long Dissolve {
             get {
                 return _dissolve;
             }
@@ -31,7 +32,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int ImageID {
+        public long ImageID {
             get {
                 return _imageID;
             }
@@ -115,7 +116,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int WatermarkID {
+        public long WatermarkID {
             get {
                 return _watermarkID;
             }
@@ -127,8 +128,5 @@ namespace SmugMug.v2.Types
                 }
             }
         }
-
-
-
     }
 }
